api/demo now takes POST request
Also corrected statuscode.

+++ xam_webapp_2022_student/oppgave_3/pages/api/demo/index.ts
@@ -7,51 +7,58 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<any>
 ) {
-  await prisma.day.deleteMany({})
-  await prisma.week.deleteMany({})
-  await prisma.lunch.deleteMany({})
-  await prisma.employee.deleteMany({})
+  if (req.method === 'POST') {
+    // Start with clean database.
+    await prisma.day.deleteMany({})
+    await prisma.week.deleteMany({})
+    await prisma.lunch.deleteMany({})
+    await prisma.employee.deleteMany({})
 
-  for (const employee of employees) {
-    await prisma.employee.create({
-      data: employee,
-    })
-  }
+    for (const employee of employees) {
+      await prisma.employee.create({
+        data: employee,
+      })
+    }
 
-  const createLunch = await prisma.lunch.create({ data: {} })
+    const createLunch = await prisma.lunch.create({ data: {} })
 
-  for (const [key, value] of Object.entries(lunch.year)) {
-    const createWeek = await prisma.week.create({
-      data: {
-        week: Number(key),
-        lunch: {
-          connect: {
-            id: createLunch.id,
+    for (const [key, value] of Object.entries(lunch.year)) {
+      const createWeek = await prisma.week.create({
+        data: {
+          week: Number(key),
+          lunch: {
+            connect: {
+              id: createLunch.id,
+            },
           },
         },
-      },
-    })
+      })
 
-    for (const [key2, value2] of Object.entries(value.week)) {
-      if (value2) {
-        await prisma.day.create({
-          data: {
-            name: key2,
-            week: {
-              connect: {
-                id: createWeek.id,
+      for (const [key2, value2] of Object.entries(value.week)) {
+        if (value2) {
+          await prisma.day.create({
+            data: {
+              name: key2,
+              week: {
+                connect: {
+                  id: createWeek.id,
+                },
+              },
+              employee: {
+                connect: {
+                  id: value2?.id,
+                },
               },
             },
-            employee: {
-              connect: {
-                id: value2?.id,
-              },
-            },
-          },
-        })
+          })
+        }
       }
     }
-  }
 
-  return res.status(200).json({ success: true })
+    return res
+      .status(201)
+      .json({ success: true, message: 'Database populated succesfully.' })
+  } else {
+    return res.status(405).json({ success: false, error: 'Method not allowed' })
+  }
 }
